renamed portfolioFilter, fixed homeServicesCard's button, added dictionary properties and translating for form
