+++ src/components/PortfolioFilter/PortfolioFilter.module.scss
@@ -1,0 +1,221 @@
+@use "@/app/variables";
+
+.filterBlock {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 24px;
+
+  @media screen and (min-width: 370px) {
+    justify-content: start;
+    column-gap: 4px;
+  }
+
+  @media screen and (min-width: variables.$tablet) {
+    column-gap: 12px;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    margin-bottom: 32px;
+  }
+}
+
+.filterItem {
+  @include variables.notification;
+  & {
+    padding: 2px 4px 2px 2px;
+    border-bottom: 2px solid transparent;
+    cursor: pointer;
+    &:hover {
+        color: variables.$hoverBtnBgColor;
+    }
+  }
+
+  @media screen and (min-width: variables.$tablet) {
+    @include variables.subtitle2-Regular;
+  }
+}
+
+.filterItemActive {
+    color: variables.$hoverBtnBgColor;
+  border-bottom: solid 1px variables.$hoverBtnBgColor;
+  }
+
+// в дальнейшем заменится на класс для активного положения фильтра
+// .filterItem:first-child {
+//   color: variables.$hoverBtnBgColor;
+//   border-bottom: solid 1px variables.$hoverBtnBgColor;
+// }
+
+.projectList {
+  display: flex;
+  flex-direction: column;
+  row-gap: 24px;
+
+  @media screen and (min-width: variables.$tablet) {
+    row-gap: 20px;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    row-gap: 40px;
+  }
+}
+
+.projectItem {
+  position: relative;
+  padding-top: 16px;
+  display: flex;
+  flex-direction: column;
+  row-gap: 12px;
+
+  @media screen and (min-width: variables.$tablet) {
+    padding-top: 12px;
+    flex-direction: row;
+    align-items: center;
+    row-gap: 0px;
+    column-gap: 16px;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    padding-top: 16px;
+    column-gap: 24px;
+  }
+}
+
+.projectItem::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  right: 0;
+  width: 100%;
+  border-top: solid 1px variables.$mainTextColor;
+
+  @media screen and (min-width: variables.$tablet) {
+    width: 47%;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    width: 588px;
+  }
+}
+
+.projectItem::after {
+  content: "";
+  position: absolute;
+  top: -8px;
+  right: 0;
+  width: 16px;
+  height: 16px;
+  background-color: variables.$hoverBtnBgColor;
+  border-radius: 50%;
+}
+
+.imgWrapper {
+  position: relative;
+  width: 100%;
+  aspect-ratio: 320/212;
+
+  @media screen and (min-width: variables.$tablet) {
+    flex: 0.689;
+    aspect-ratio: 463/309;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    flex: 0 1 auto; // обнуление flex
+    width: 894px;
+    height: 596px;
+  }
+}
+
+.img {
+  display: block;
+  object-fit: cover;
+}
+
+.contentWrapper {
+  @media screen and (min-width: variables.$tablet) {
+    flex: 0.311;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    flex: 0 1 auto; // обнуление flex
+  }
+}
+
+.type {
+  @include variables.subtitle2-Regular;
+  & {
+    margin-bottom: 4px;
+  }
+
+  @media screen and (min-width: variables.$tablet) {
+    margin-bottom: 8px;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    margin-bottom: 12px;
+  }
+}
+
+.location {
+  margin-bottom: 2px;
+
+  @media screen and (min-width: variables.$tablet) {
+    margin-bottom: 4px;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    margin-bottom: 8px;
+  }
+}
+
+.area {
+  margin-bottom: 2px;
+
+  @media screen and (min-width: variables.$tablet) {
+    margin-bottom: 4px;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    margin-bottom: 8px;
+  }
+}
+
+.style {
+  margin-bottom: 2px;
+
+  @media screen and (min-width: variables.$tablet) {
+    margin-bottom: 4px;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    margin-bottom: 8px;
+  }
+}
+
+.year {
+  margin-bottom: 8px;
+
+  @media screen and (min-width: variables.$tablet) {
+    margin-bottom: 16px;
+  }
+
+  @media screen and (min-width: variables.$desktop) {
+    margin-bottom: 24px;
+  }
+}
+
+
+// todo code to filter
+
+// .filterItem {
+//     border-bottom: 2px solid transparent;
+//     &:hover {
+//         color: variables.$hoverBtnBgColor;
+//     }
+//   }
+  
+//   .active {
+//     color: variables.$hoverBtnBgColor;
+//   border-bottom: solid 1px variables.$hoverBtnBgColor;
+//   }
+
